Move methods from config companion object to object Config
