+++ zio-config/src/main/scala/zio/config/Config.scala
@@ -1,5 +1,7 @@
 package zio.config
 
+import java.net.URI
+
 sealed trait Config[A] {
   self =>
   final def zip[B](that: => Config[B]): Config[(A, B)] = Config.Zip(self, that)
@@ -63,4 +65,16 @@ object Config {
 
   def collectAll[A](configList: List[Config[A]]): Config[List[A]] =
     sequence(configList)
+
+  def int(path: String): Config[Int]               = Config.Source(path, PropertyType.IntType)
+  def double(path: String): Config[Double]         = Config.Source(path, PropertyType.DoubleType)
+  def string(path: String): Config[String]         = Config.Source(path, PropertyType.StringType)
+  def long(path: String): Config[Long]             = Config.Source(path, PropertyType.LongType)
+  def short(path: String): Config[Short]           = Config.Source(path, PropertyType.ShortType)
+  def uri(path: String): Config[URI]               = Config.Source(path, PropertyType.UriType)
+  def bigDecimal(path: String): Config[BigDecimal] = Config.Source(path, PropertyType.BigDecimalType)
+  def float(path: String): Config[Float]           = Config.Source(path, PropertyType.FloatType)
+  def bigInt(path: String): Config[BigInt]         = Config.Source(path, PropertyType.BigIntType)
+  def boolean(path: String): Config[Boolean]       = Config.Source(path, PropertyType.BooleanType)
+  def byte(path: String): Config[Byte]             = Config.Source(path, PropertyType.ByteType)
 }
