+++ zio-config/src/test/scala/zio/config/ListAndOptionalTests.scala
@@ -10,7 +10,7 @@ object ListAndOptionalTests extends Properties("List and options tests") with Te
   private val genOverallConfig: Gen[OverallConfig] =
     Gen.option(genId).map(t => OverallConfig(t.map(t => Option(Option(t)))))
 
-  val cId: Config[Id] = string("kId").xmap(Id)(_.value)
+  val cId: Config[Id] = Config.string("kId").xmap(Id)(_.value)
 
   val cOverallConfig: Config[OverallConfig] =
     cId.optional.optional.optional.xmap(OverallConfig)(_.option)
